requirements.txt
google-genai
streamlit google-genai
